+++ pst.h
@@ -11,9 +11,9 @@
 
 // constants
 
-constexpr int_fast32_t Opening = 0;
-constexpr int_fast32_t Endgame = 1;
-constexpr int_fast32_t StageNb = 2;
+constexpr int_fast8_t Opening = 0;
+constexpr int_fast8_t Endgame = 1;
+constexpr int_fast8_t StageNb = 2;
 
 // macros
